Fix rtmpdump.c compiler warning
rtmpdump.c: In function `GetLastKeyframe':
rtmpdump.c:297:7: warning: unknown conversion type character `l' in format
rtmpdump.c:297:7: warning: too many arguments for format

http://stackoverflow.com/q/586928/how-should-i-print-types-like-off-t-and-size-t

+++ rtmpdump.c
@@ -293,8 +293,8 @@ GetLastKeyframe(FILE * file,	// output file [in]
 
   bAudioOnly = (dataType & 0x4) && !(dataType & 0x1);
 
-  RTMP_Log(RTMP_LOGDEBUG, "bAudioOnly: %d, size: %llu", bAudioOnly,
-      (unsigned long long) size);
+  RTMP_Log(RTMP_LOGDEBUG, "bAudioOnly: %d, size: %lu", bAudioOnly,
+      (unsigned long) size);
 
   // ok, we have to get the timestamp of the last keyframe (only keyframes are seekable) / last audio frame (audio only streams)
